# api/openapi.yaml
openapi: 3.1.0
info:
  title: ProxyLoom Target Management API
  version: 0.1.0-m0
  summary: Target contract for the complete ProxyLoom self-hosted subscription manager.
  description: |
    This is the complete target contract, not evidence that every operation is running.
    The current implemented surface is described by openapi-runtime.yaml.
    All timestamps use RFC 3339. Resource IDs are lowercase UUIDv7 strings.
    Cookie-authenticated unsafe requests also require X-CSRF-Token.
    List cursors are opaque and expire after 15 minutes.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: /
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/setup/admin:
    post:
      tags: [System]
      operationId: initializeAdministrator
      summary: Create the first administrator exactly once.
      security: []
      parameters:
        - $ref: '#/components/parameters/SetupToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSetupRequest'
      responses:
        '201':
          description: Administrator initialized and session created.
          headers:
            Set-Cookie:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/session:
    post:
      tags: [Authentication]
      operationId: createSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: Authenticated session.
          headers:
            Set-Cookie:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      tags: [Authentication]
      operationId: getSession
      responses:
        '200':
          description: Current session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Authentication]
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Session revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/session/reauthenticate:
    post:
      tags: [Authentication]
      operationId: reauthenticateSession
      summary: Extend the five-minute recent-authentication window for secret operations.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password]
              properties:
                password: {type: string, minLength: 1, maxLength: 1024, writeOnly: true}
      responses:
        '200':
          description: Updated session and recent-authentication deadline.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Session'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/session/password:
    post:
      tags: [Authentication]
      operationId: changeAdministratorPassword
      summary: Change the current administrator password and revoke all sessions.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [current_password, new_password]
              properties:
                current_password: {type: string, minLength: 1, maxLength: 1024, writeOnly: true}
                new_password: {type: string, minLength: 1, maxLength: 1024, writeOnly: true}
      responses:
        '204':
          description: Password changed and all sessions revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/v1/sources:
    get:
      tags: [Sources]
      operationId: listSources
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema:
            type: string
            enum: [unknown, healthy, degraded, unhealthy, disabled]
        - name: query
          in: query
          schema: {type: string, maxLength: 200}
      responses:
        '200':
          description: Stable page ordered by updated_at desc then id desc.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourcePage'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Sources]
      operationId: createSource
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceCreate'
      responses:
        '201':
          description: Source with its first draft revision.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/sources/{sourceId}:
    parameters:
      - $ref: '#/components/parameters/SourceId'
    get:
      tags: [Sources]
      operationId: getSource
      responses:
        '200':
          description: Source details with secrets masked.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Sources]
      operationId: updateSource
      description: Creates a new immutable draft revision.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/SourcePatch'
      responses:
        '200':
          description: Updated source and draft revision.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Sources]
      operationId: archiveSource
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Source archived; referenced history is retained.
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
...
  /api/v1/nodes:
    get:
      tags: [Nodes]
      operationId: listNodes
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: source_id
          in: query
          schema: {$ref: '#/components/schemas/Id'}
        - name: output_id
          in: query
          schema: {$ref: '#/components/schemas/Id'}
        - name: protocol
          in: query
          schema: {type: string}
        - name: health
          in: query
          schema:
            $ref: '#/components/schemas/NodeHealthState'
        - name: query
          in: query
          schema: {type: string, maxLength: 200}
      responses:
        '200':
          description: Stable node occurrence page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodePage'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/node-checks:
    post:
      tags: [Nodes]
      operationId: checkNodeScope
      summary: Queue a source, output or all-node manual check with global Health Key deduplication.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BulkNodeCheckRequest'}
      responses:
        '202':
          $ref: '#/components/responses/JobAccepted'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/collections:
    get:
      tags: [Resources]
      operationId: listCollections
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Collection page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionedResourcePage'
    post:
      tags: [Resources]
      operationId: createCollection
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
      responses:
        '201':
          $ref: '#/components/responses/RevisionedResourceCreated'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/pipelines:
    get:
      tags: [Resources]
      operationId: listPipelines
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Pipeline page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionedResourcePage'
    post:
      tags: [Resources]
      operationId: createPipeline
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineCreate'
      responses:
        '201':
          $ref: '#/components/responses/RevisionedResourceCreated'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/v1/jobs:
    get:
      tags: [Jobs]
      operationId: listJobs
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: type
          in: query
          schema: {$ref: '#/components/schemas/JobType'}
        - name: status
          in: query
          schema: {$ref: '#/components/schemas/JobStatus'}
      responses:
        '200':
          description: Jobs ordered by created_at desc then id desc.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobPage'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: proxyloom_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  parameters:
    SetupToken:
      name: X-ProxyLoom-Setup-Token
      in: header
      required: true
      description: One-time 256-bit token generated locally by proxyloom bootstrap-token; expires in 30 minutes.
      schema: {type: string, minLength: 43, maxLength: 128}
...
  schemas:
...
    AdminSetupRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {type: string, minLength: 3, maxLength: 64}
        password: {type: string, minLength: 1, maxLength: 1024, writeOnly: true}
        timezone: {type: string, default: Asia/Shanghai}
...
    Session:
      type: object
      additionalProperties: false
      required: [administrator, csrf_token, expires_at, recent_auth_until]
      properties:
        administrator:
          type: object
          required: [id, username]
          properties:
            id: {$ref: '#/components/schemas/Id'}
            username: {type: string}
        csrf_token: {type: string, readOnly: true}
        expires_at: {$ref: '#/components/schemas/Timestamp'}
        recent_auth_until: {$ref: '#/components/schemas/Timestamp'}
...